+++ tests/auth/expiring.py
@@ -103,6 +103,7 @@ class TestCredentialsOnline:
         c = Credentials(app_env[0], app_env[1], asynchronous=True)
         token = await c.request_client_token()
         assert token.refresh_token is None
+        await c.close()
 
     def test_refresh_user_token(self, app_env, user_refresh):
         c = Credentials(app_env[0], app_env[1])
@@ -114,6 +115,7 @@ class TestCredentialsOnline:
         c = Credentials(app_env[0], app_env[1], asynchronous=True)
         token = await c.refresh_user_token(user_refresh)
         assert token.refresh_token is not None
+        await c.close()
 
     def test_bad_arguments_raises_error(self):
         c = Credentials('id', 'secret')